backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
redis==5.0.1
pydantic==1.8.2
python-multipart==0.0.5
websockets==10.0
motor==3.3.2
pymongo==4.6.1